# master/server/docs/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/deployment": {
      "get": {
        "description": "list all deployment",
        "produces": [
          "application/json"
        ],
        "tags": [
          "deployment"
        ],
        "operationId": "listDeployment",
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Deployment"
              }
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      },
      "post": {
        "description": "create deployment",
        "produces": [
          "application/json"
        ],
        "tags": [
          "deployment"
        ],
        "operationId": "createDeployment",
        "parameters": [
          {
            "name": "DeploymentReq",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeploymentReq"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      }
    },
    "/deployment/{id}": {
      "put": {
        "description": "update deployment",
        "produces": [
          "application/json"
        ],
        "tags": [
          "deployment"
        ],
        "operationId": "updateDeployment",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "name": "DeploymentReq",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DeploymentReq"
            }
          }
        ],
        "responses": {
          "204": {
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      },
      "delete": {
        "description": "delete deployment",
        "produces": [
          "application/json"
        ],
        "tags": [
          "deployment"
        ],
        "operationId": "deleteDeployment",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      }
    },
    "/storage_item": {
      "get": {
        "description": "List StorageItem",
        "produces": [
          "application/json"
        ],
        "tags": [
          "storageItem"
        ],
        "operationId": "listStorageItem",
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/StorageItem"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      },
      "post": {
        "description": "Create StorageItem",
        "produces": [
          "application/json"
        ],
        "tags": [
          "storageItem"
        ],
        "operationId": "createStorageItem",
        "parameters": [
          {
            "type": "string",
            "name": "name",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "name": "type",
            "in": "formData",
            "required": true
          },
          {
            "type": "file",
            "name": "file",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/StorageItem"
            }
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      }
    },
    "/storage_item/exist": {
      "post": {
        "description": "Create Exist StorageItem",
        "produces": [
          "application/json"
        ],
        "tags": [
          "storageItem"
        ],
        "operationId": "createExistStorageItem",
        "parameters": [
          {
            "name": "storageItemExistReq",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StorageItemExistReq"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/StorageItem"
            }
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      }
    },
    "/storage_item/{id}": {
      "delete": {
        "description": "Delete StorageItem",
        "produces": [
          "application/json"
        ],
        "tags": [
          "storageItem"
        ],
        "operationId": "deleteStorageItem",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/StorageItem"
            }
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      }
    },
    "/storage_item/{type}": {
      "get": {
        "description": "List StorageItem By Type",
        "produces": [
          "application/json"
        ],
        "tags": [
          "storageItem"
        ],
        "operationId": "listStorageItemByType",
        "parameters": [
          {
            "type": "string",
            "name": "type",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/StorageItem"
            }
          },
          "403": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Deployment": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "x-go-name": "Content",
          "example": "123"
        },
        "id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID",
          "example": 1
        },
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "test-image"
        }
      },
      "x-go-package": "github.com/Ch4r1l3/cFuzz/master/server/models"
    },
    "DeploymentReq": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "content": {
          "type": "string",
          "x-go-name": "Content",
          "example": "123"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "test-image"
        }
      },
      "x-go-package": "github.com/Ch4r1l3/cFuzz/master/server/controller"
    },
    "ErrResp": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error",
          "example": "some error"
        }
      },
      "x-go-package": "github.com/Ch4r1l3/cFuzz/utils"
    },
    "StorageItem": {
      "description": "item that record some file info",
      "type": "object",
      "properties": {
        "existsInImage": {
          "type": "boolean",
          "x-go-name": "ExistsInImage",
          "example": true
        },
        "id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID",
          "example": 1
        },
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "afl"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type",
          "example": "fuzzer"
        }
      },
      "x-go-package": "github.com/Ch4r1l3/cFuzz/master/server/models"
    },
    "StorageItemExistReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "afl"
        },
        "path": {
          "type": "string",
          "x-go-name": "Path",
          "example": "/tmp/afl/123"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type",
          "example": "fuzzer"
        }
      },
      "x-go-package": "github.com/Ch4r1l3/cFuzz/master/server/controller"
    },
...
  }
}